move non encrypt and decrypt methods to shiftable module
