+++ lib/encrypt.rb
@@ -1,0 +1,1 @@
+
