--- lib/shifter.rb
@@ -1,7 +1,0 @@
-class Shifter
-  attr_reader :alphabet
-
-  def initialize
-    @alphabet = ("a".."z").to_a << " "
-  end
-end
