--- test/encrypt_test.rb
@@ -1,1 +1,0 @@
-
